refactor: item > items로파일 이름 변경
