+++ week2-shopping-mall/pages/home/scripts/index.js
@@ -1,7 +1,7 @@
 import { attachEventListeners } from "./eventHandlers.js";
 import { attachCommonEventListener } from "../../../common/index.js";
 import { displayItems } from "./utils.js";
-import { SHOPPING_LIST } from "../../../constants/item.js";
+import { SHOPPING_LIST } from "../../../constants/items.js";
 
 document.addEventListener("DOMContentLoaded", (event) => {
   displayItems(SHOPPING_LIST);
